Invalidate request cache multiple times on action
Unfortunately, actions do not always block until they are fully
complete and things (for example responses from ncip) may take a while
to show up in the audit log, etc.  so we have to expire the cache a few
times to keep the record reasonably up to date as that happens.